refactor(chain): separate finalized VM inventory model
